Restructure CMake config for Catch2 testing
- I learned a lot from this exercise.
- I should only have one project in the root
CMakeLists.txt. Having a project(test) was creating a
separate build directory in test/.
- I struggled to link the path_finder files
in src to test. The trick in the end was to define
path_finder.cpp and grid.cpp as a library SEARCH,
which I could think link both the MAIN and TEST
executables against.
- You need to link the library and also include
the headers. Not entirely sure about why this is
at this point.